MySQL_Advanced

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -2,8 +2,9 @@
 DELIMITER //
 CREATE PROCEDURE ComputeAverageScoreForUser(IN userId INT)
 BEGIN
-DECLARE average INT;
+DECLARE average FLOAT;
 SELECT sum(score) / count(score) INTO average FROM corrections WHERE user_id = userId;
+
 UPDATE users
 SET average_score = average
 WHERE id = userId;
